+++ temp_util.py
@@ -24,9 +24,9 @@ CN_NEWS_FOLDER_NAME = "news"
 FINAL_VIDEOS_FOLDER_NAME = "final_videos"
 EVENING_TAG = "_E"
 EVENING = False
-CHINADAILY = 'chinadaily'
-CHINADAILY_EN = 'chinadaily_en'
-CHINADAILY_HK = 'chinadaily_hk'
+CHINADAILY = 'cn_daily'
+CHINADAILY_EN = 'cn_daily_en'
+CHINADAILY_HK = 'cn_daily_hk'
 ALJ = 'alj'
 CFR = 'cfr'
 BBC = 'bbc'
